Take a fast-paced multiplayer game approach
Instead of emitting several events I give a try with fast-paced update with a picture of the whole game. The idea is to keep everyone with the same game and take advantage of cache and avoid extra processing time. For that matter I've added a "dirty" status in the gameState object.